Move render engine into SyntaxKit; inject Subprocess backend via closure [skip ci]
Relocate Runner, RunInput, and RunError from skit into SyntaxKit/Execution so
the DSL-render engine is usable as a library, not locked to the CLI. Runner no
longer depends on Subprocess: it takes an injected `run` closure
(SwiftInvocation -> ProcessResult) and skit supplies the Subprocess-backed
implementation.

skit-side Subprocess seam, factored into extensions:
- TerminationStatus.exitCode — collapse to one Int32 (128 + signal convention).
- Subprocess.Configuration.runSwift(for:) — build the swift config, run it, and
  map the record into a ProcessResult (the `run` closure handed to Runner).
- Skit+Run no longer hand-rolls the spawn; it passes the extension as the backend.

SyntaxKit-side cleanup:
- New SwiftInvocation value type (libPath + wrappedPath) replaces the generic
  `Configuration` name that clashed with Subprocess.Configuration; own file.
- SwiftRunOutcome simplified to `.completed(ProcessResult)`; dead Foundation
  import dropped.
- Split directory/batch mode into Runner+Directory.swift, bringing Runner.swift
  back under the file-length limit; removed leftover commented code and fixed
  spacing.

Build clean; all 398 tests pass (incl. SkitSubprocessTimeoutTests).
