+++ Sources/SyntaxKit/Execution/SwiftInvocation.swift
@@ -1,5 +1,5 @@
 //
-//  RunError.swift
+//  SwiftInvocation.swift
 //  SyntaxKit
 //
 //  Created by Leo Dion.
@@ -27,22 +27,11 @@
 //  OTHER DEALINGS IN THE SOFTWARE.
 //
 
-#if canImport(Subprocess)
-
-  /// Typed error surfaced by `Runner`. It decouples the renderer from the CLI:
-  /// `Runner` reports *what* went wrong, and `Skit.Run.run` decides the process
-  /// exit code (so the engine can also be driven in-process from a library).
-  internal enum RunError: Error {
-    /// The input path was invalid — missing, or a directory given without `-o`.
-    case invalidInput(String)
-    /// Single-file mode: the spawned `swift` exited non-zero. Carries that code
-    /// (e.g. a compile failure, `124` on timeout, `128 + signal`).
-    case renderFailed(exitCode: Int32)
-    /// Directory mode: at least one input failed; the batch exit code is `1`.
-    case batchFailed
-    /// A wrapped Foundation/Subprocess failure (file read/write, spawn error)
-    /// that has no dedicated exit-code mapping.
-    case unexpected(any Error)
-  }
-
-#endif
+/// The inputs needed to invoke `swift` on one wrapped DSL program: the lib
+/// directory (for `-I`/`-L`/`-rpath`) and the path to the wrapped source.
+/// `Runner` hands this to its `run` closure, which performs the actual spawn —
+/// the seam that keeps the engine free of any Subprocess dependency.
+package struct SwiftInvocation: Sendable {
+  package let libPath: String
+  package let wrappedPath: String
+}
